強化學習專論/final project/final_project_env/requirements.txt
bidict==0.22.1
blinker==1.6.3
click==8.1.7
cloudpickle==3.0.0
dataclasses-json==0.6.1
Farama-Notifications==0.0.4
Flask==3.0.0
Flask-SocketIO==5.3.6
gymnasium==0.28.1
h11==0.14.0
itsdangerous==2.1.2
jax-jumpy==1.0.0
Jinja2==3.1.2
MarkupSafe==2.1.3
marshmallow==3.20.1
mypy-extensions==1.0.0
nptyping==1.4.4
numpy==1.25.2
opencv-python==4.8.1.78
packaging==23.2
PettingZoo==1.22.3
Pillow==10.1.0
pybullet==3.1.7
python-engineio==4.8.0
python-socketio==5.10.0
PyYAML==6.0.1
scipy==1.11.3
simple-websocket==1.0.0
typing-inspect==0.9.0
typing_extensions==4.8.0
typish==1.9.3
Werkzeug==3.0.1
wsproto==1.2.0
yamldataclassconfig==1.5.0

setuptools==65.5.1
requests==2.31.0
matplotlib==3.8.0